requirements.txt
adal==1.2.7
aiofiles==23.2.1
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
anytree==2.12.1
appdirs==1.4.4
argcomplete==3.3.0
asgiref==3.8.1
asttokens 
attrs==23.2.0
autograd==1.6.2
azure-ai-ml==1.16.0
azure-common==1.1.28
azure-core==1.30.1
azure-graphrbac==0.61.1
azure-identity==1.16.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.4.0
azure-mgmt-keyvault==10.3.0
azure-mgmt-network==25.3.0
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.1.0
azure-ml==0.0.1
azure-search==1.0.0b2
azure-search-documents==11.5.0
azure-storage-blob==12.20.0
azure-storage-file-datalake==12.15.0
azure-storage-file-share==12.16.0
azureml==0.2.7
azureml-core==1.56.0
azureml-dataprep==5.1.6
azureml-dataprep-native==41.0.0
azureml-dataprep-rslex==2.22.2
azureml-fsspec==1.3.1
azureml-rag==0.2.32
backoff==2.2.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.1.3
beartype==0.18.5
beautifulsoup4==4.12.3
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.3
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
coloredlogs==15.0.1
comm 
ConfigArgParse==1.7
contextlib2==21.6.0
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
dataclasses-json==0.6.6
datasets==2.19.1
datashaper==0.0.49
debugpy 
decorator 
Deprecated==1.2.14
deprecation==2.1.0
devtools==0.12.2
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
easydict==1.13
email_validator==2.2.0
environs==11.0.0
exceptiongroup 
executing 
faiss-cpu==1.8.0.post1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.4.0
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2023.10.0
future==1.0.0
gensim==4.3.2
google-api-core==2.19.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
gradio==4.40.0
gradio_client==1.2.0
graspologic==3.4.1
graspologic-native==1.2.1
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
humanfriendly==10.0
hyppo==0.4.0
idna==3.7
importlib_metadata 
importlib_resources==6.4.0
ipykernel 
ipython 
isodate==0.6.1
jedi 
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpickle==3.0.4
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client 
jupyter_core 
kiwisolver==1.4.5
knack==0.11.0
kubernetes==30.1.0
lancedb==0.10.1
langchain==0.1.20
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith==0.1.98
llama-index==0.10.38
llama-index-agent-openai==0.2.5
llama-index-cli==0.1.12
llama-index-core==0.10.42
llama-index-embeddings-langchain==0.1.2
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-anyscale==0.1.4
llama-index-llms-azure-openai==0.1.10
llama-index-llms-langchain==0.1.4
llama-index-llms-openai==0.1.20
llama-index-multi-modal-llms-openai==0.1.6
llama-index-packs-raptor==0.1.3
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.23
llama-index-readers-llama-parse==0.1.4
llama-index-readers-smart-pdf-loader==0.1.4
llama-parse==0.4.3
llamaindex-py-client==0.1.19
llmsherpa==0.1.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
matplotlib-inline 
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
ndg-httpsclient==0.5.1
nest_asyncio 
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.40.0
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.3
overrides==7.7.0
packaging==23.2
pandas==2.2.2
paramiko==3.4.0
parso 
pathspec==0.12.1
patsy==0.5.6
pexpect 
pickleshare 
pillow==10.3.0
pkginfo==1.10.0
platformdirs 
portalocker==2.8.2
posthog==3.5.0
POT==0.9.4
prompt-toolkit 
proto-plus==1.23.0
protobuf==4.25.3
psutil 
ptyprocess 
pure-eval 
py==1.11.0
pyaml-env==1.2.1
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic
pydantic_core
pydash==8.0.1
pydub==0.25.1
Pygments 
PyJWT==2.8.0
pylance==0.14.1
PyNaCl==1.5.0
pynndescent==0.5.13
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pysbd==0.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq 
ragas==0.1.8
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
retry==0.9.2
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruff==0.5.6
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.12.0
seaborn==0.13.2
SecretStorage==3.3.3
semantic-version==2.10.0
sentence-transformers==2.7.0
setuptools==69.5.1
shellingham==1.5.4
simplejson==3.19.2
six 
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
stack-data 
starlette==0.37.2
statsmodels==0.14.2
strictyaml==1.7.3
striprtf==0.0.26
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken
tokenizers==0.19.1
tomlkit==0.12.0
torch==2.3.0
tornado 
tqdm==4.66.4
traitlets 
transformers==4.41.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions 
tzdata==2024.1
ujson==5.10.0
umap-learn==0.5.6
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth 
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp 